requirements.txt
flask
mysql-connector-python
numpy
scikit-fuzzy